Various improvements

+++ index.php
@@ -1,13 +1,15 @@
 <?php
+header("Content-Type: text/html; charset=ISO-8859-1");
+
 include "inc.php";
 ?>
 <!doctype html>
 <html>
 <head>
-<title>Pastetenfabrik</title>
+<title>Pastetenfabrik at <?=$_SERVER["SERVER_NAME"] ?></title>
 </head>
 <body>
-<h1>Pastetenfabrik</h1>
+<h1>Pastetenfabrik at <?=$_SERVER["SERVER_NAME"] ?></h1>
 
 <p>Current time: <i><?php echo date('r'); ?></i></p>
 
@@ -27,13 +29,13 @@ if (isset($_POST['do'])) {
 			echo '<p><a href="'.$id.'.html"><b>Your Paste: '.$id.'</b></a></p>';
 		}
 	}
-	else if ($_POST['do'] == 'del') {
+	else if ($_POST['do'] == 'del' && isset($_POST['pw'])) {
 		include "pastes.php";
 
 		if (isset($p[$_POST['id']])) {
 			$u = resolve(sha512($_POST['pw']));
 			
-			if ($u == 0 || $u == $p[$_POST['id']][2]) {
+			if ($u == 0 || $u == $p[$_POST['id']][2] || ($public && $u != -1)) {
 				unlink('p/'.$_POST['id'].'.txt.bz2');
 				file_put_contents('pastes.php','unset($p["'.$_POST['id'].'"]); // '.date('r').' - USER'.$u."\n",FILE_APPEND);
 				echo '<p><a href="'.$_POST['id'].'.html"><b>'.$_POST['id'].'</b></a> deleted.</p>';
@@ -65,9 +67,7 @@ if (isset($_POST['do'])) {
 <form action="index.php" method="post">
 	<input type="hidden" name="do" value="del" />
 	ID: <input type="text" name="id" /><br />
-	<?php
-		echo 'Password: <input type="password" name="pw" /><br />';
-	?>
+	Password: <input type="password" name="pw" /><br />
 	<input type="submit" value="Submit" />
 </form>
 </fieldset>
